check if database process already terminated

+++ libs/database/server/databaseserver.cpp
@@ -545,7 +545,7 @@ void DatabaseServer::stopDatabaseProcess()
 
     d->databaseProcess->terminate();
 
-    if (!d->databaseProcess->waitForFinished(10000))
+    if (d->databaseProcess->state() == QProcess::Running && !d->databaseProcess->waitForFinished(10000))
     {
         qCDebug(DIGIKAM_DATABASESERVER_LOG) << "Database process will be killed now";
         d->databaseProcess->kill();
